# tooth-pick/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ToothPick API",
    "description": "API completa para la plataforma dental B2B/B2C ToothPick",
    "version": "1.0.0",
    "contact": {
      "name": "ToothPick Team",
      "email": "support@toothpick.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor de desarrollo"
    },
    {
      "url": "https://toothpick.vercel.app/api",
      "description": "Servidor de producción"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number" },
          "category": { "type": "string" },
          "brand": { "type": "string" },
          "stock": { "type": "integer" },
          "isActive": { "type": "boolean" }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/OrderItem" } },
          "total": { "type": "number" },
          "status": { 
            "type": "string", 
            "enum": ["pending", "confirmed", "shipped", "delivered", "cancelled"] 
          },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "productId": { "type": "string" },
          "quantity": { "type": "integer" },
          "price": { "type": "number" }
        }
      },
...
    }
  },
  "paths": {
...
    "/products": {
      "get": {
        "summary": "Listar productos",
        "tags": ["Productos"],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filtrar por categoría"
          },
          {
            "name": "brand",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filtrar por marca"
          },
          {
            "name": "search",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Búsqueda de texto"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de productos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Product" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear producto",
        "tags": ["Productos"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Product" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Producto creado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Product" }
              }
            }
          },
          "400": { "description": "Datos inválidos" },
          "401": { "description": "No autorizado" },
          "403": { "description": "Solo providers pueden crear productos" }
        }
      }
    },
    "/orders": {
      "get": {
        "summary": "Listar pedidos",
        "tags": ["Pedidos"],
        "responses": {
          "200": {
            "description": "Lista de pedidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Order" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear pedido",
        "tags": ["Pedidos"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "items": {
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/OrderItem" }
                  },
                  "shippingAddress": { "type": "object" },
                  "paymentMethod": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pedido creado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Order" }
              }
            }
          }
        }
      }
    },
...
  },
...
}